backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
pyyaml==6.0.2
pandas==2.2.3
rapidfuzz==3.10.0
python-dotenv==1.0.1
openai==1.51.0
python-multipart==0.0.12
pdfplumber==0.11.4
python-docx==1.1.2
Pillow==10.4.0
boto3==1.35.0
